file_processing/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
tiktoken
pandas
numpy
langchain
langchain-openai
openai
python-dotenv
azure-ai-formrecognizer
azure-cognitiveservices-speech
azure-ai-textanalytics
azure-search-documents==11.4.0b8
azure-storage-blob
azure-identity